requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
streamlit>=1.28.0
plotly>=5.18.0
requests>=2.31.0
python-telegram-bot>=20.6
schedule>=1.2.0
textblob>=0.17.1
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
